Add files via upload
Changed speed_up so that it works as a toggle. Did the same for the scope. 

Do you think that we could get all the asteroid generation functions to be within the asteroid class itself? I want to make a power-up object that inherits those methods from the asteroid class so that we can have power-ups spawn randomly, collide with the player etc.